+++ neptune/new/internal/backends/hosted_neptune_backend.py
@@ -18,7 +18,7 @@ import os
 import platform
 import re
 import uuid
-from typing import List, Optional, Dict, Iterable
+from typing import List, Optional, Dict, Iterable, Tuple, Any
 
 import click
 import urllib3
@@ -609,7 +609,7 @@ class HostedNeptuneBackend(NeptuneBackend):
             raise FetchAttributeNotFoundException(path_to_str(path))
 
     @with_api_exceptions_handler
-    def fetch_atom_attribute_values(self, run_uuid: uuid.UUID, path: List[str]) -> Namespace:
+    def fetch_atom_attribute_values(self, run_uuid: uuid.UUID, path: List[str]) -> List[Tuple[str, AttributeType, Any]]:
         params = {
             'experimentId': str(run_uuid),
         }
@@ -619,12 +619,10 @@ class HostedNeptuneBackend(NeptuneBackend):
                 # don't want to catch "ns/attribute/other" while looking for "ns/attr"
                 namespace_prefix += "/"
             result = self.leaderboard_client.api.getExperimentAttributes(**params).response().result
-            attributes = [attr for attr in result.attributes if attr.name.startswith(namespace_prefix)]
-            run_struct = RunStructure()
-            for attr in attributes:
-                value = map_attribute_result_to_value(attr)
-                run_struct.set(parse_path(attr.name), (attr.type, value))
-            return run_struct.get_structure()
+            return [
+                (attr.name, attr.type, map_attribute_result_to_value(attr))
+                for attr in result.attributes if attr.name.startswith(namespace_prefix)
+            ]
         except HTTPNotFound:
             raise RunUUIDNotFound(run_uuid)